test(nodeRuntime): platform-stubbed specs so Windows CI keeps branch coverage
The skipIf(win32) shim tests left every ensureNodeRuntime branch uncovered
on Windows runners, dropping global branch coverage to 89.13% (<90%).
Move them to ts/nodeRuntime.spec.ts with process.platform stubbed per
test: POSIX runners now also exercise the win32 early-return and Windows
runners exercise the full shim path (fs ops all work there; the shim is
simply never consulted by real spawns). Adds a liveEnv mutation test.
