+++ ts/serve.spec.ts
@@ -1,8 +1,6 @@
-import { mkdtemp, readFile, rm, stat } from "fs/promises";
-import { tmpdir } from "os";
-import path from "path";
-import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { ensureNodeRuntime, isNoNodeExecError, oxmgrVersionHasWindowsFix } from "./serve.ts";
+import { readFile } from "fs/promises";
+import { describe, expect, it } from "vitest";
+import { isNoNodeExecError, oxmgrVersionHasWindowsFix } from "./serve.ts";
 
 // Guards the Windows daemon-manager selection: on Windows we only PREFER oxmgr
 // when the installed build carries the daemon-socket-inheritance fix. Stock
@@ -32,82 +30,6 @@ describe("oxmgrVersionHasWindowsFix", () => {
   it("returns false on unparseable output", () => {
     expect(oxmgrVersionHasWindowsFix("")).toBe(false);
     expect(oxmgrVersionHasWindowsFix("oxmgr (unknown)")).toBe(false);
-  });
-});
-
-// The bun-only-no-node path (sym003): pm2/oxmgr bins are `#!/usr/bin/env node`
-// scripts, so with no node on PATH they die at the shebang. ensureNodeRuntime
-// must bridge that with a node→bun shim in ay's own bin dir.
-describe("ensureNodeRuntime", () => {
-  let home: string;
-  let savedPath: string | undefined;
-  let savedHome: string | undefined;
-
-  beforeEach(async () => {
-    home = await mkdtemp(path.join(tmpdir(), "ay-node-shim-"));
-    savedPath = process.env.PATH;
-    savedHome = process.env.AGENT_YES_HOME;
-    process.env.AGENT_YES_HOME = home;
-  });
-
-  afterEach(async () => {
-    process.env.PATH = savedPath;
-    if (savedHome === undefined) delete process.env.AGENT_YES_HOME;
-    else process.env.AGENT_YES_HOME = savedHome;
-    await rm(home, { recursive: true, force: true });
-  });
-
-  const noNode = (cmd: string) => (cmd === "bun" ? "/fake/bin/bun" : null);
-
-  it.skipIf(process.platform === "win32")(
-    "writes an executable node→bun shim and prepends it to PATH when node is absent",
-    async () => {
-      const shim = await ensureNodeRuntime(noNode);
-      expect(shim).toBe(path.join(home, "bin", "node"));
-      const body = await readFile(shim!, "utf-8");
-      expect(body).toBe("#!/bin/sh\nexec '/fake/bin/bun' \"$@\"\n");
-      expect((await stat(shim!)).mode & 0o111).not.toBe(0); // executable
-      expect(process.env.PATH!.split(path.delimiter)[0]).toBe(path.join(home, "bin"));
-    },
-  );
-
-  it.skipIf(process.platform === "win32")("does not duplicate the PATH entry", async () => {
-    await ensureNodeRuntime(noNode);
-    await ensureNodeRuntime(noNode);
-    const binDir = path.join(home, "bin");
-    const hits = process.env.PATH!.split(path.delimiter).filter((p) => p === binDir);
-    expect(hits).toHaveLength(1);
-  });
-
-  it.skipIf(process.platform === "win32")(
-    "leaves an up-to-date shim untouched on repeat calls (read-only paths stay read-only)",
-    async () => {
-      const shim = (await ensureNodeRuntime(noNode))!;
-      const before = await stat(shim);
-      await new Promise((r) => setTimeout(r, 20));
-      await ensureNodeRuntime(noNode);
-      const after = await stat(shim);
-      expect(after.mtimeMs).toBe(before.mtimeMs); // not rewritten
-    },
-  );
-
-  it.skipIf(process.platform === "win32")(
-    "single-quotes a bun path containing sh-special characters",
-    async () => {
-      const shim = await ensureNodeRuntime((cmd) =>
-        cmd === "bun" ? "/opt/we$ird `dir'/bun" : null,
-      );
-      const body = await readFile(shim!, "utf-8");
-      expect(body).toBe("#!/bin/sh\nexec '/opt/we$ird `dir'\\''/bun' \"$@\"\n");
-    },
-  );
-
-  it("is a no-op when a real node exists", async () => {
-    expect(await ensureNodeRuntime(() => "/usr/bin/whatever")).toBeNull();
-  });
-
-  it("is a no-op when bun is missing too (nothing to shim to)", async () => {
-    expect(await ensureNodeRuntime(() => null)).toBeNull();
   });
 });
